+++ packages/std/src/errors/std_error.rs
@@ -453,6 +453,7 @@ pub enum OverflowOperation {
     Mul,
     Pow,
     Shr,
+    Shl,
 }
 
 impl fmt::Display for OverflowOperation {